Correct @id return value from r_navigation
The `@id` key in the response from r_navigation is actually the url path to the current navigation request. Here I have changed the return for `@id` to the objectId and moved the url path that was in `@id` to a new `dts:references` key. This seems to make more sense to me but perhaps it doesn't conform to the standard.